+++ src/main/java/carracing/domain/RandomScoreStrategy.java
@@ -1,4 +1,4 @@
-package car_racing.domain;
+package carracing.domain;
 
 public class RandomScoreStrategy implements ScoreStrategy {
     @Override
