+++ src/main/java/carracing/domain/ScoreStrategy.java
@@ -1,4 +1,4 @@
-package car_racing.domain;
+package carracing.domain;
 
 public interface ScoreStrategy {
     Integer generateScore();
