+++ src/main/java/carracing/domain/StringParser.java
@@ -1,4 +1,4 @@
-package car_racing.domain;
+package carracing.domain;
 
 public class StringParser {
     public static String[] parse(String inputString) {
