+++ src/main/java/stringcalculator/Calculator.java
@@ -1,4 +1,4 @@
-package string_calculator;
+package stringcalculator;
 
 import java.util.Arrays;
