+++ src/main/java/stringcalculator/OutputView.java
@@ -1,4 +1,4 @@
-package string_calculator;
+package stringcalculator;
 
 public class OutputView {
     public static void printResult(Integer result) {
