+++ src/main/java/stringcalculator/StringParser.java
@@ -1,4 +1,4 @@
-package string_calculator;
+package stringcalculator;
 
 import java.util.ArrayList;
 import java.util.Arrays;
