+++ src/test/java/stringcalculator/StringParserTest.java
@@ -1,4 +1,4 @@
-package string_calculator;
+package stringcalculator;
 
 import org.assertj.core.api.Assertions;
 import org.junit.jupiter.params.ParameterizedTest;
